% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polys.R
\name{adjust_poly}
\alias{adjust_poly}
\title{Adjusts edges to become a valid poly}
\usage{
adjust_poly(edges, delta, weights = NULL)
}
\value{
a vector of edges adjusted to become valid
}
\description{
Adjusts edges to become a valid poly
}
\examples{
adjust_geom(c(1,2,4), 1)
adjust_geom(c(1,2,4), 1, weights = c(1,1,2))
}
