% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polys.R
\name{create_polys}
\alias{create_polys}
\title{Partitions a set of edges into a list of polys with n sides (doesn't exist)}
\usage{
create_polys(df_edges, n = 3)
}
\arguments{
\item{df_edges}{a dataframe with columns `node1`, `node2`, `edge`, `weight`. weight
col can be excluded. First two columns must be node1, node2 because of use
of igraph:::graph_from_data_frame.}

\item{n}{the number of nodes to include in each shape.}
}
\value{
a list of vectors, where each vector contains the row indices from
  the input df_edges that make up a poly
}
\description{
I haven't figured out how to do this, I am using the igraph package and creating 
triangles in `get_triangles` instead.
}
