% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangle.R
\name{get_triangles}
\alias{get_triangles}
\title{Find all triangles in a set of nodes and edges}
\usage{
get_triangles(df_edges)
}
\arguments{
\item{df_edges}{a dataframe with columns `node1`, `node2`, `edge`, `weight`. weight
col can be excluded. First two columns must be node1, node2 because of use
of igraph:::graph_from_data_frame.}
}
\value{
a list of lists, where each element is the row indices from
  df_edges that make up a triangle.
}
\description{
Find all triangles in a set of nodes and edges
}
\examples{
df_edges <-
tidyr::tribble(
 ~node1, ~node2, ~edge,
 "a", "b", 1,
 "a", "c", 1,
 "b", "c", 3
)

tris <- get_triangles(df_edges)
}
