% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polys.R
\name{invalid_polys}
\alias{invalid_polys}
\title{Checks polygon inequality rule, if invalid returns excess distance value}
\usage{
invalid_polys(df_edges, polys)
}
\arguments{
\item{df_edges}{a dataframe with columns `node1`, `node2`, `edge`, `weight`. weight
col can be excluded. First two columns must be node1, node2 because of use
of igraph:::graph_from_data_frame.}

\item{polys}{a list of lists, where each element is the row indices from
df_edges that make up a polygon}
}
\value{
vector of same length as polys containing 0 for valid polys or the
  excess distance
}
\description{
For any polygon the longest side has to be shorter than the sum of all other
sides. This will return that distance, returns 0 for valid polygons.
}
\examples{
df_edges <-
tidyr::tribble(
 ~node1, ~node2, ~edge,
 "a", "b", 1,
 "a", "c", 1,
 "b", "c", 3
)

tris <- get_triangles(df_edges)
invalid_polys(df_edges, tris)
}
