% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{triangle_plot}
\alias{triangle_plot}
\title{Create plot of triangles}
\usage{
triangle_plot(df_edges, polys, tolerance = 0.001)
}
\arguments{
\item{df_edges}{a dataframe with columns `node1`, `node2`, `edge`, `weight`. weight
col can be excluded. First two columns must be node1, node2 because of use
of igraph:::graph_from_data_frame.}

\item{polys}{a list of lists, where each element is the row indices from
df_edges that make up a polygon}

\item{tolerance}{allowed error in checking polygon inequality}
}
\value{
a plot
}
\description{
Distance between nodes is not scaled to distance values - it can't be since
we want to show invalid triangles! Invalid triangles are highlighted in
red.Since edges are shared, a valid triangle is one where at least one side
is grey.
}
\examples{
df_edges <-
tidyr::tribble(
 ~node1, ~node2, ~edge,
 "a", "b", 1,
 "a", "c", 1,
 "b", "c", 3
)

tris <- get_triangles(df_edges)
triangle_plot(df_edges, tris)
}
